pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-surveillance"
version = "0.1.0"
description = "Trade surveillance backend: feature pipelines, anomaly scoring, and investigation orchestration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "psycopg[binary]>=3.1",
    "scikit-learn>=1.3",
    "sqlalchemy>=2.0",
    "shap>=0.43",
    "langgraph>=0.2",
    "langchain>=0.2",
    "langchain-anthropic>=0.1",
    "anthropic>=0.25",
    "python-dotenv>=1.0",
    "boto3>=1.28",
    "tqdm>=4.65",
    "typing-extensions>=4.8",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["."]
include = ["trade_surveillance*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["trade_surveillance", "tests"]
